Updates handleRequestInsertClipping to allow for shadowdom
When we get the active element, if that element is within shadowdom, we must get the active element within the shadow dom. This fix checks to see if the shadowRoot exists as a property of the element and then updates activeElt to the activeElement in the shadow root.

Tested in firefox using both shadow dom and non shadow dom.